Update: edit mode
- Add: After the path is edited, animation now correctly works - streets that aren't connected to any previous once are skipped, vehicles that are on streets that aren't connected or are closed don't move
+ doesnt matter if the vehicle gets on the street from start or from end
- Add: functions to get point distance, random edited path length
- Fix: input file coordinates are now correctly connected

+++ src/main/java/map/Coordinate.java
@@ -38,12 +38,20 @@ public class Coordinate
 		return this.y;
 	}
 	
-	public void setX(float x)
+	/**
+	 * Umoznuje nastavi souradnici X
+	 * @param x
+	 */
+	public void setX(double x)
 	{
 		this.x = x;
 	}
 	
-	public void setY(float y)
+	/**
+	 * Umoznuje nastavi souradnici Y
+	 * @param y
+	 */
+	public void setY(double y)
 	{
 		this.y = y;
 	}
